requirements.txt
cachetools==2.0.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==0.12.2
google-api-python-client==1.6.6
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.11.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
neat-python==0.92
oauth2client==4.1.2
oauthlib==2.0.7
pyasn1==0.4.2
pyasn1-modules==0.2.1
pygame==1.9.4
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.14.1
